Put part modifiers in their own files
And normalize the interface.  It's somewhat large unfortunately but need to cover every scenario.
