+++ B9PartSwitch/B9PartSwitch.csproj
@@ -102,6 +102,18 @@
     <Compile Include="PartSwitch\AttachNodeMover.cs" />
     <Compile Include="PartSwitch\AttachNodeModifierInfo.cs" />
     <Compile Include="PartSwitch\ModuleB9PartInfo.cs" />
+    <Compile Include="PartSwitch\PartModifiers\IPartModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartAttachNodeModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCenterOfBuoyancyModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCenterOfDisplacementModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCoLOffsetModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCoMOffsetModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCoPOffsetModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartCrashToleranceModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartMaxTempModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartModifierBase.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartSkinMaxTempModifier.cs" />
+    <Compile Include="PartSwitch\PartModifiers\PartStackSymmetryModifier.cs" />
     <Compile Include="PartSwitch\PartSwitchFlightDialog.cs" />
     <Compile Include="PartSwitch\TextureReplacement.cs" />
     <Compile Include="PartSwitch\TextureSwitchInfo.cs" />
