+++ B9PartSwitch/PartSwitch/ModuleB9PartInfo.cs
@@ -1,5 +1,6 @@
 ﻿using UniLinq;
 using System.Collections.Generic;
+using B9PartSwitch.PartSwitch.PartModifiers;
 
 namespace B9PartSwitch
 {
